Reference

Contents

Reference

Contents

Index

Index

TulipaClustering.ClusteringData

TulipaClustering.DemandData

TulipaClustering.GenerationAvailabilityData

TulipaClustering.read_clustering_data_from_csv_folder

TulipaClustering.read_csv_with_schema

TulipaClustering.ClusteringData — Type

Structure to hold all parts of an energy problem.

TulipaClustering.DemandData — Type

Schema for the demand.csv file.

TulipaClustering.GenerationAvailabilityData — Type

Schema for the generation-availability.csv file.

TulipaClustering.read_clustering_data_from_csv_folder — Method

read_clustering_data_from_csv_folder(input_folder)

Returns the TulipaClustering.ClusteringData reading all data from CSV files in the input_folder.

The following files are expected to exist in the input folder:

demand.csv: Following the TulipaClustering.DemandData specification.

generation-availability.csv: Following the TulipaClustering.GenerationAvailabilityData specification.

The output contains:

demand: a DataFrame of demand values at different nodes per time step

generation_availability: a DataFrame of availability coefficients for different generation technologies located at different nodes per time step

TulipaClustering.read_csv_with_schema — Method

read_csv_with_schema(file_path, schema)

Reads the csv with file_name at location path validating the data using the schema. It is assumes that the file's header is at the second row. The first row of the file contains some metadata information that is not used.